requirements.txt
langchain
python-dotenv
pypdf
chromadb
langchain_google_genai
langchain-community
uvicorn
fastapi
python-multipart